API Areas

+++ database/migrations/2020_12_22_134935_create_areas_table.php
@@ -17,7 +17,7 @@ class CreateAreasTable extends Migration
             $table->increments('id');
             $table->integer('category_id')->unsigned()->nullable(); // tipo de Category
             $table->string('name', 255)->nullable();
-            $table->integer('status')->nullable();
+            $table->TINYINT('status')->nullable();
             $table->string('category')->nullable();
             $table->timestamps();
         });
